requirements.txt
Flask
flask-cors
google-cloud-storage
google-auth
vertexai
gunicorn